Alternative running version of IC
To be able to better run the IC locally, a copy of the ImmortalCoordinator was
implemented. This IC can now be started locally and enables TTD, too.
It was necessary to move LongPair out of AmbrosiaLib into the
SharedAmbrosiaTools to remove the circular dependency.

The checkpointing is initialized with the IC program and is run for each message.

Missing:
[] A way to restart TTD without needing to recalculate all information of the
checkpointing strategy.
[] Information about the current processed event and other dynamic information.