Feat: Included splitting of log records
To enable the IC to handle the events separately, batched log records need to
be splitted into their separate messages. This is done by adapting the code
used by the language binding and splits all rpc-batches into single rpcs.
The rpcs are sent separately to the LB. Before each action it is determined
whether a checkpoint should be made or not (might change in future).