Store additional information about record-processing into checkpoints
As the log record messages are splitted into separate entries, the reprocessing
should not run a message twice. To prevent this, the offset and the last
processed sub-message are stored into the sub-checkpoints. To enable full
checkpoints, this information should also be included into full checkpoints as
well. It should be kept in mind that this should not break previous checkpoints.

Furthermore: Tested with batched and realized that currently only the first
message of log records is used.
-> Fix: Use all messages within a log record (and within batched messages).