Feat: Enable extraction of events
To be able to extract events for further processing, the corresponding method
was implemented. It may not be necessary to call this method at this point, but
it is possible.

Further: Apply abstraction of Event type to AmbrosiaEvent. And implemented a
count of processed events to keep track of the "internal" timestamp.