Feat: First possibility to run checkpointing on normal execution
To run on normal execution, another lock is required, leading to totally
sequential processing of the messages. As the messages need to be split (as
again Batches should be handled by the IC), this could take a while.
A problem could be, that the high watermarks are committed too early and thus
recovery might be broken with this commit. This should be fixed in a later
commit!!!

This commit enables the checkpointing on normal run and works after each single
RPC as well as after a specified size.